+++ tests_integration/__tests__/early-exit.js
@@ -3,6 +3,9 @@
 const prettier = require("../../tests_config/require_prettier");
 const runPrettier = require("../runPrettier");
 const constant = require("../../src/cli/constant");
+const util = require("../../src/cli/util");
+const commonUtil = require("../../src/common/util");
+const getSupportInfo = require("../../src/common/support").getSupportInfo;
 
 describe("show version with --version", () => {
   runPrettier("cli/with-shebang", ["--version"]).test({
@@ -23,20 +26,35 @@ describe(`show detailed usage with --help l (alias)`, () => {
   });
 });
 
-constant.detailedOptions.forEach(option => {
-  const optionNames = [
-    option.description ? option.name : null,
-    option.oppositeDescription ? `no-${option.name}` : null
-  ].filter(Boolean);
+commonUtil
+  .arrayify(
+    Object.assign(
+      {},
+      util.createDetailedOptionMap(
+        getSupportInfo(null, {
+          showDeprecated: true,
+          showUnreleased: true,
+          showInternal: true
+        }).options
+      ),
+      util.normalizeDetailedOptionMap(constant.options)
+    ),
+    "name"
+  )
+  .forEach(option => {
+    const optionNames = [
+      option.description ? option.name : null,
+      option.oppositeDescription ? `no-${option.name}` : null
+    ].filter(Boolean);
 
-  optionNames.forEach(optionName => {
-    describe(`show detailed usage with --help ${optionName}`, () => {
-      runPrettier("cli", ["--help", optionName]).test({
-        status: 0
+    optionNames.forEach(optionName => {
+      describe(`show detailed usage with --help ${optionName}`, () => {
+        runPrettier("cli", ["--help", optionName]).test({
+          status: 0
+        });
       });
     });
   });
-});
 
 describe("show warning with --help not-found", () => {
   runPrettier("cli", ["--help", "not-found"]).test({
